+++ CHANGELOG.md
@@ -4,6 +4,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.8.1] - 2023-06-12
+### :bug: Bug Fixes
+- [`cb9cac1`](https://github.com/requarks/changelog-action/commit/cb9cac16822feb7033a12db5731511e82f106d5c) - handle related PR issues query failures *(commit by [@NGPixel](https://github.com/NGPixel))*
+
+
+## [v1.8.0] - 2023-04-18
+### :sparkles: New Features
+- [`a67af14`](https://github.com/requarks/changelog-action/commit/a67af14034e62802f8a8fa856a763a76d925df0d) - add support for GHE *(PR [#25](https://github.com/requarks/changelog-action/pull/25) by [@anden-dev](https://github.com/anden-dev))*
+  - :arrow_lower_right: *addresses issue [#24](undefined) opened by [@anden-dev](https://github.com/anden-dev)*
+
+
+## [v1.7.0] - 2023-03-17
+### :sparkles: New Features
+- [`01c1b24`](https://github.com/requarks/changelog-action/commit/01c1b24b234e079288271046481d408baad64656) - add reverseOrder option to list commits from newer to older *(commit by [@NGPixel](https://github.com/NGPixel))*
+
+### :bug: Bug Fixes
+- [`94af0c3`](https://github.com/requarks/changelog-action/commit/94af0c3dfeae6180da49e87ec06a24880614c081) - handle types in uppercase *(commit by [@NGPixel](https://github.com/NGPixel))*
+- [`d47b63a`](https://github.com/requarks/changelog-action/commit/d47b63a7f846dd6c4aa803c597a12d413121fd59) - handle commits with no author info *(commit by [@NGPixel](https://github.com/NGPixel))*
+
+
+## [v1.6.0] - 2022-12-15
+### :sparkles: New Features
+- [`f64e045`](https://github.com/requarks/changelog-action/commit/f64e045b5e7d73289888b92aa7cf6b9c8443f497) - include referenced issues from PRs *(commit by [@NGPixel](https://github.com/NGPixel))*
+
+
 ## [v1.5.0] - 2022-11-15
 ### :sparkles: New Features
 - [`0192e0e`](https://github.com/requarks/changelog-action/commit/0192e0ed0553ee53648e187d784ccfdefe9e16b3) - add includeInvalidCommits option *(commit by [@NGPixel](https://github.com/NGPixel))*
@@ -70,3 +95,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [v1.3.2]: https://github.com/requarks/changelog-action/compare/v1.3.1...v1.3.2
 [v1.4.0]: https://github.com/requarks/changelog-action/compare/v1.3.2...v1.4.0
 [v1.5.0]: https://github.com/requarks/changelog-action/compare/v1.4.0...v1.5.0
+[v1.6.0]: https://github.com/requarks/changelog-action/compare/v1.5.0...v1.6.0
+[v1.7.0]: https://github.com/requarks/changelog-action/compare/v1.6.0...v1.7.0
+[v1.8.0]: https://github.com/requarks/changelog-action/compare/v1.7.0...v1.8.0
+[v1.8.1]: https://github.com/requarks/changelog-action/compare/v1.8.0...v1.8.1